Working avatar select form
